backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain-community==0.0.27
langchain-core==0.1.30
python-dotenv==1.0.1
mysql-connector-python==8.3.0 